+++ modules/timeline-layers/src/index.ts
@@ -1,0 +1,8 @@
+// deck.gl-community
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export type {HorizonGraphLayerProps} from './layers/horizon-graph-layer/horizon-graph-layer';
+export {HorizonGraphLayer} from './layers/horizon-graph-layer/horizon-graph-layer';
+export type {MultiHorizonGraphLayerProps} from './layers/horizon-graph-layer/multi-horizon-graph-layer';
+export {MultiHorizonGraphLayer} from './layers/horizon-graph-layer/multi-horizon-graph-layer';